🚑 chore: api내 모델 분리 - room, dashboard
