--- server/api/templates/rooms_list.html
@@ -1,36 +1,0 @@
-<!-- templates/rooms_list.html -->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Room List</title>
-</head>
-<body>
-    <h1>Room List</h1>
-    <ul>
-        {% for room in rooms %}
-            <li>Name: {{ room.name }}, Type: {{ room.type }}, Difficulty: {{ room.difficulty }}</li>
-        {% empty %}
-            <li>No rooms available.</li>
-        {% endfor %}
-    </ul>
-
-    <!-- Create Room Form -->
-    <form action="{% url 'create_room' %}" method="post">
-        {% csrf_token %}
-        <input type="text" name="name" placeholder="Room Name" required>
-        <input type="text" name="type" placeholder="Room Type" required>
-        <input type="text" name="difficulty" placeholder="Difficulty" required>
-        <button type="submit">Create Room</button>
-    </form>
-
-    <!-- Create Match Form -->
-    <form action="{% url 'create_match' %}" method="post">
-        {% csrf_token %}
-        <input type="text" name="user1_uuid" placeholder="User 1 UUID" required>
-        <input type="text" name="user2_uuid" placeholder="User 2 UUID" required>
-        <input type="text" name="win_user" placeholder="Winning User UUID" required>
-        <button type="submit">Create Match</button>
-    </form>
-</body>
-</html>
